Fix callbacks a little
Note: events are propagated, so when we a country 
is inserted, the before-insert on form will capture
both regions and countries, while the after-insert
will stop propagation once captured.

So I made sure we have after and before insert
on both countries and regions to be able to correctly
stop. Also: the event-handler must be on a container
of the link triggering the action. 

Also: as items are inserted dynamically we listen
for events on $(document) and use a filter.